+++ src/interfaces/vacancy.ts
